feat: add support for overloading `html-react-parser` options (#62)
* feat:  Add doc to overload html-react-parser options

* fix: format

* fix: formating in overloading-wp-behaviour

* fix: add condition to use only one config at a time

* Implement codesplitting and use types from package

* docs: update doc and include file paths

* docs: add changeset

* refactor: change sequnce of blocks

* docs: improve flow between html-react-parser and blocks

* docs: update config-api with `parserOptions` reference

---------
