+++ .changeset/sixty-hotels-stare.md
@@ -1,0 +1,6 @@
+---
+"@snapwp/core": minor
+"@snapwp/next": minor
+---
+
+feat: Add support for overloading html-react-parser options
