+++ docs/overloading-wp-behavior.md
@@ -4,9 +4,116 @@ SnapWP takes an additive approach to headless WordPress development, where WordP
 
 This tutorial explains how to overload default WordPress behavior using SnapWP, focusing on customizing blocks, creating custom frontend routes, and `html-react-parser` component mapping.
 
+## Default HTML Rendering with `Parse` and [`html-react-parser`](https://www.npmjs.com/package/html-react-parser)
+
+By default, SnapWP uses WordPress as the source of truth for both content _and_ markup. The `Parse` component in SnapWP allows you to convert raw HTML strings into React components, replacing specific elements with custom Next.js components.
+
+This component can be used anywhere in your Next.js application to progressively enhance the rendering of HTML content.
+
+Here's a basic example:
+
+```tsx
+import React from 'react';
+import { Parse } from '@snapwp/next';
+import { type BlockData } from '@snapwp/core';
+
+/**
+ * Renders the default block.
+ * @param props - The props for the component.
+ * @param props.renderedHtml - The rendered HTML.
+ * @return The rendered default block.
+ */
+export default function Default( { renderedHtml }: BlockData ) {
+	return <Parse html={ renderedHtml || '' } />;
+}
+```
+
+### Overloading the HTML-to-React component mapping
+
+SnapWP allows you to extend and overload the default `HTMLReactParserOptions` used by the `Parse` component. This is useful for customizing the rendering of specific HTML elements, such as images, links, or other custom elements.
+
+### 1. Creating a custom `HTMLReactParserOptions` object`
+
+Create a new Custom Component to overload the [default html-react-parser options](../packages/next/src/react-parser/options.tsx).
+
+```tsx
+// MyCustomReactParser.tsx
+
+import React from 'react';
+import {
+	Element,
+	type DOMNode,
+	type HTMLReactParserOptions,
+} from 'html-react-parser';
+import { CustomImage } from '@/CustomImage'; // custom image component defined in your project
+/**
+ *  Using typescript type guard to check if a DOMNode is an Element.
+ * Ref : https://github.com/remarkablemark/html-react-parser/issues/221#issuecomment-784073240
+ * @param domNode The DOM element
+ * @return parser options
+ */
+const isElement = ( domNode: DOMNode ): domNode is Element => {
+	const isTag = domNode.type === 'tag';
+	const hasAttributes = ( domNode as Element ).attribs !== undefined;
+
+	return isTag && hasAttributes;
+};
+
+export const customParserOptions: HTMLReactParserOptions = {
+	replace: ( domNode ) => {
+		if ( isElement( domNode ) ) {
+			const { attribs, children, name, type } = domNode;
+			const { class: className, style, ...attributes } = attribs;
+			const { href } = attribs;
+
+			if ( type === 'tag' && name === 'img' ) {
+				return (
+					<CustomImage
+						{ ...attributes }
+						src={ attribs.src }
+						alt={ attribs.alt || '' }
+						height={ height }
+						width={ width }
+						className={ className }
+						style={ styleObject }
+						image={ imageAttributes }
+						//adding custom id
+						id="overriding id"
+					/>
+				);
+			}
+
+			return undefined;
+		}
+
+		return undefined;
+	},
+};
+```
+
+### 2. Pass customParserOptions to overload
+
+```tsx
+// snapwp.config.ts
+
+import type { SnapWPConfig } from '@snapwp/core/config';
+import { customParserOptions } from './src/MyCustomReactParser';
+
+const config: SnapWPConfig = {
+	/* passing custom options to overload default react-parser options  */
+	parserOptions: customParserOptions,
+};
+
+export default config;
+```
+
+---
+
 ## Overloading Blocks
 
-SnapWP allows you to customize the rendering of individual WordPress blocks. This is useful for modifying the default appearance or behavior of a block to better suit your design.
+In addition to overloading the underling rendering of default HTML content, SnapWP allows you to target the rendering of individual WordPress blocks.
+
+The `@snapwp/blocks` package provides a large and growing number of [Block Components](https://github.com/rtCamp/snapwp/blob/develop/packages/blocks/src/blocks/index.ts), which you can in turn overload with custom components or unregister to fallback to rendering with [html-react-parser](#1-creating-a-custom-htmlreactparseroptions-object).
 
 > [!TIP]
 > When using WordPress [Block Themes](https://wordpress.org/documentation/article/block-themes/), pretty much _everything_ is a block.
@@ -65,7 +172,10 @@ export default function Page() {
 }
 ```
 
-Now, whenever a `core/paragraph` block is encountered, your `MyCustomParagraph` component will be used to render it. Any other blocks will use the default rendering unless you provide a custom component in `blockDefinitions`. Setting the value to `null` will use the [default block rendering](#default-block-rendering).
+Now, whenever a `core/paragraph` block is encountered, your `MyCustomParagraph` component will be used to render it. Any other blocks will use the default rendering unless you provide a custom component in `blockDefinitions`.
+
+> [!TIP]
+> Setting the block definition value to `null` will use the [default block rendering](#default-html-rendering-with-parse-and-html-react-parser).
 
 ---
 
@@ -178,27 +288,3 @@ export default function Page() {
 
 -   **Local Styles**: Use CSS Modules (as shown with `styles.module.css`) or any other CSS-in-JS solution for component-specific styling.
 -   **Global and Theme Styles**: Classes like `wp-block-heading`, `has-text-align-center`, and `has-x-large-font-size` (likely from your WordPress `theme.json` and/or global CSS) are automatically available.
-
-## Default Block Rendering
-
-This tutorial covers how SnapWP uses the `Parse` component to render HTML content. The `Parse` component converts HTML strings into React components, which is essential for displaying default WordPress blocks.
-
-### Using `Parse`
-
-The `Parse` component takes an HTML string (`renderedHtml`) and converts it into React components. Here's a basic example:
-
-```tsx
-import React from 'react';
-import { Parse } from '@snapwp/next';
-import { type BlockData } from '@snapwp/core';
-
-/**
- * Renders the default block.
- * @param props - The props for the component.
- * @param props.renderedHtml - The rendered HTML.
- * @return The rendered default block.
- */
-export default function Default( { renderedHtml }: BlockData ) {
-	return <Parse html={ renderedHtml || '' } />;
-}
-```
